Refactor Appearance, added documentation, rebuild migration
