+++ Characters.Server/Characters.Server.csproj
@@ -106,6 +106,7 @@
     <Reference Include="System.ComponentModel.DataAnnotations">
       <Private>False</Private>
     </Reference>
+    <Reference Include="System.Data" />
     <Reference Include="Z.EntityFramework.Plus.EF6, Version=1.10.1.0, Culture=neutral, PublicKeyToken=59b66d028979105b, processorArchitecture=MSIL">
       <HintPath>..\packages\Z.EntityFramework.Plus.EF6.1.10.1\lib\net45\Z.EntityFramework.Plus.EF6.dll</HintPath>
       <Private>False</Private>
@@ -115,17 +116,15 @@
     <Compile Include="CharacterManager.cs" />
     <Compile Include="Events\CharacterEventArgs.cs" />
     <Compile Include="Events\CharacterSessionEventArgs.cs" />
-    <Compile Include="Migrations\201903230015160_Init.cs" />
-    <Compile Include="Migrations\201904122043530_AddSessionFk.cs" />
-    <Compile Include="Migrations\201908030133090_AddCharacterInventory.cs" />
+    <Compile Include="Migrations\201908170055570_Init.cs" />
     <Compile Include="Migrations\Configuration.cs" />
-    <Compile Include="Models\Appearance.cs" />
     <Compile Include="Models\Character.cs" />
-    <Compile Include="Models\Apparel.cs" />
+    <Compile Include="Models\CharacterFacialTrait.cs" />
+    <Compile Include="Models\CharacterHeritage.cs" />
     <Compile Include="Models\CharacterInventory.cs" />
     <Compile Include="Models\CharacterSession.cs" />
-    <Compile Include="Models\FaceShape.cs" />
-    <Compile Include="Models\Heritage.cs" />
+    <Compile Include="Models\CharacterStyle.cs" />
+    <Compile Include="Models\CharacterTrait.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="CharactersController.cs" />
     <Compile Include="Storage\StorageContext.cs" />
