+++ Characters.Server/Storage/StorageContext.cs
@@ -8,8 +8,6 @@ namespace IgiCore.Characters.Server.Storage
 	{
 		public DbSet<Character> Characters { get; set; }
 
-		public DbSet<Apparel> Appearances { get; set; }
-
 		public DbSet<CharacterInventory> CharacterInventories { get; set; }
 
 		public DbSet<CharacterSession> CharacterSessions { get; set; }
