--- Characters.Shared/Models/Apparel/Component.cs
@@ -1,9 +1,0 @@
-namespace IgiCore.Characters.Shared.Models.Apparel
-{
-	public class Component
-	{
-		public ComponentTypes Type { get; set; }
-		public int Index { get; set; }
-		public int Texture { get; set; }
-	}
-}
