--- Characters.Shared/Models/Apparel/Prop.cs
@@ -1,9 +1,0 @@
-namespace IgiCore.Characters.Shared.Models.Apparel
-{
-	public class Prop
-	{
-		public PropTypes Type { get; set; }
-		public int Index { get; set; }
-		public int Texture { get; set; }
-	}
-}
