--- Characters.Shared/Models/Appearance/FeatureTypes.cs
@@ -1,17 +1,0 @@
-namespace IgiCore.Characters.Shared.Models.Appearance
-{
-	public enum FeatureTypes
-	{
-		Beard = 1,
-		Eyebrows = 2,
-		Age = 3,
-		Makeup = 4,
-		Blush = 5,
-		Complexion = 6,
-		SunDamage = 7,
-		Lipstick = 8,
-		MolesAndFreckles = 9,
-		Chest = 10,
-		Blemishes = 11
-	}
-}
